# swagger/tag_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "tag_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Tags"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/tag": {
      "get": {
        "operationId": "Tags_list",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTagListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Tags"
        ]
      },
      "post": {
        "operationId": "Tags_put",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTagResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiTagLite"
            }
          }
        ],
        "tags": [
          "Tags"
        ]
      }
    },
    "/api/v1/tag/search/{like}": {
      "get": {
        "operationId": "Tags_search",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTagListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "like",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Tags"
        ]
      }
    },
    "/api/v1/tag/{id}": {
      "get": {
        "operationId": "Tags_get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTagResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Tags"
        ]
      }
    }
  },
  "definitions": {
    "apiDebug": {
      "type": "object",
      "properties": {
        "messages": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apiError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "boolean"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "msgDetails": {
          "type": "string"
        }
      }
    },
    "apiTagFull": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "isDeleted": {
          "type": "boolean"
        }
      }
    },
    "apiTagListResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiTagLite"
          }
        },
        "error": {
          "$ref": "#/definitions/apiError"
        },
        "debug": {
          "$ref": "#/definitions/apiDebug"
        }
      }
    },
    "apiTagLite": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "isDeleted": {
          "type": "boolean"
        }
      }
    },
    "apiTagResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/apiTagFull"
        },
        "error": {
          "$ref": "#/definitions/apiError"
        },
        "debug": {
          "$ref": "#/definitions/apiDebug"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}